+++ InstaConnect.Business/Services/PostLikeService.cs
@@ -4,6 +4,7 @@ using InstaConnect.Business.Abstraction.Services;
 using InstaConnect.Business.Models.DTOs.PostLike;
 using InstaConnect.Business.Models.Results;
 using InstaConnect.Business.Models.Utilities;
+using InstaConnect.Data.Abstraction.Helpers;
 using InstaConnect.Data.Abstraction.Repositories;
 using InstaConnect.Data.Models.Entities;
 
@@ -16,19 +17,22 @@ namespace InstaConnect.Business.Services
         private readonly IPostLikeRepository _postLikeRepository;
         private readonly IPostRepository _postRepository;
         private readonly IUserRepository _userRepository;
+        private readonly IAccountManager _accountManager;
 
         public PostLikeService(
             IMapper mapper,
             IResultFactory resultFactory,
             IPostLikeRepository postLikeRepository,
             IPostRepository postRepository,
-            IUserRepository userRepository)
+            IUserRepository userRepository,
+            IAccountManager accountManager)
         {
             _mapper = mapper;
             _resultFactory = resultFactory;
             _postLikeRepository = postLikeRepository;
             _postRepository = postRepository;
             _userRepository = userRepository;
+            _accountManager = accountManager;
         }
 
         public async Task<IResult<ICollection<PostLikeResultDTO>>> GetAllAsync(
@@ -83,8 +87,17 @@ namespace InstaConnect.Business.Services
             return okResult;
         }
 
-        public async Task<IResult<PostLikeResultDTO>> AddAsync(PostLikeAddDTO postLikeAddDTO)
+        public async Task<IResult<PostLikeResultDTO>> AddAsync(string userId, PostLikeAddDTO postLikeAddDTO)
         {
+            var validUser = _accountManager.ValidateUser(userId, postLikeAddDTO.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostLikeResultDTO>();
+
+                return forbiddenResult;
+            }
+
             var existingUser = await _userRepository.FindEntityAsync(f => f.Id == postLikeAddDTO.UserId);
 
             if (existingUser == null)
@@ -140,13 +153,22 @@ namespace InstaConnect.Business.Services
 
         public async Task<IResult<PostLikeResultDTO>> DeleteAsync(string userId, string id)
         {
-            var existingPostLike = await _postLikeRepository.FindEntityAsync(pl => pl.Id == id && pl.UserId == userId);
+            var existingPostLike = await _postLikeRepository.FindEntityAsync(pl => pl.Id == id);
 
             if (existingPostLike == null)
             {
                 var notFoundResult = _resultFactory.GetNotFoundResult<PostLikeResultDTO>();
 
                 return notFoundResult;
+            }
+
+            var validUser = _accountManager.ValidateUser(userId, existingPostLike.UserId);
+
+            if (!validUser)
+            {
+                var forbiddenResult = _resultFactory.GetForbiddenResult<PostLikeResultDTO>();
+
+                return forbiddenResult;
             }
 
             await _postLikeRepository.DeleteAsync(existingPostLike);
